For me, jackson 2.8.11.1 couldn't be found in https://repo.maven.apache.org/maven2, so I downgraded it to 2.8.11.
I changed the group IDs of the example artifacts to com.graphql-java-kickstart from
com.graphql-java to match the main artifacts

I added slf4j to the feature. This now seems to be required.

There now seems to be an incompatibility between the karaf maven plugin and the latest
maven-install-plugin:3.0.0-M1. See https://stackoverflow.com/a/52622275/518627. So I
put in specific references to the previous plugin versions in order to fix this.

The change that *actually* makes the result work is to change
the dependency on javax-servlet-api to be 3.1.0 in the build.gradle
file. Otherwise the graphql-java-servlet depends on javax-servlet-api
version 4, whereas the Pax Web Http whiteboard depends on 3.1 and so
never recognises the graphql-java-servlet component.