Improve schema OSGi whiteboard update performance by batching
Allow the ability to delay the update of the schema through a config
option. By default this is 0, leaving the behaviour unchanged, and no
penalty is paid.
If the delay is set to a non zero value, it only updates the schema if no
changes to the whiteboard have happened within that delay timeout.

Also fixed up a couple of things with the whiteboard. Using
policy=DYNAMIC is the normal approach, the component can activate
and gets references added as appropriate. The "typesProviders" method
was in error as the name didn't match the unbind. I believe that what
was happenning before is that the component was being activated
and deactivated each time a new reference was added. Now it's activated
once and the whiteboard works correctly.