ci: build and publish javax flavour
The 'release' github action has a new job to build the javax flavour of
this project. This job has three main steps:
 - Replace all the jakarta imports and dependencies with the javax ones
 - Downgrade springframework dependency to the version 5.*
 - Using jdk 11 to build the project and publish it

The cold migration from jakarta to javax is done via a new bash script
`.github/replaceJakartaWithJavax.sh` that uses mainly the sed command to
perform the replacement operations.

In gradle.properties the SOURCE_COMPATIBILITY and the TARGET_COMPATIBILITY
constants are downgraded to 11 in order to allow the users to use the
latest version of this library even in projects where the jdk11 is mandatory.

Updated the README with a brief explanation of the difference between
the two flavours of this project.